implement build_all argument

+++ doc/source/conf.py
@@ -10,6 +10,7 @@ import subprocess
 from sphinx.application import Sphinx
 import os
 import shutil
+import warnings
 
 project = 'Light Installation'
 copyright = '2024, erichstuder'
@@ -42,10 +43,13 @@ html_static_path = ['_static']
 def pre_build_copy_textual_documentation(app):
     src = os.path.join(app.srcdir, '../../Textual/doc/build')
     dest = os.path.join(app.srcdir, '_static/auto_copied/Textual/build')
-    if os.path.exists(dest):
-        shutil.rmtree(dest)
-    shutil.copytree(src, dest)
-    print(f"Copied {src} to {dest}")
+    if os.path.exists(src):
+        if os.path.exists(dest):
+            shutil.rmtree(dest)
+        shutil.copytree(src, dest)
+        print(f"Copied Textual documentation from {src} to {dest}")
+    else:
+        warnings.warn(f"No Textual documentation to copy.")
 
 def setup(app):
     app.connect('builder-inited', pre_build_copy_textual_documentation)
